Indexed mesh rendering with some bare minimum abstractions.
Vertex attributes are stored in a single binding in an interleaved buffer (array of structures) when submitting the buffer to the GPU.

The original mesh on CPU side is stored in an opposite format - structure of arrays - currently supporting positions, uvs, normals, colors.

This allows to make the buffers optional and the buffer size flexible, where we only allocate the buffer for number of bytes we are using, while also keeping the vertex attribute data cache-friendly on GPU side.

(the code can use some cleanup - todo next)